website/docs/r/cognito_user_pool.markdown: Document attribute constraints requirement
There is a known issue with the `aws_cognito_user_pool` resource that causes a recreation if schema attributes of type `String` or `Number` are used without an attribute constraints configuration block. This change does not fix the issue, but updates the documentation for the resource with a note indicating the requirements to prevent future invocations of the said issue.