Set aws_eip.allocation_id on read
allocation_id on schema is marked as Computed, but never set to a value. It is
also undocumented, and is likely legacy, since it would duplicate the Id field.
The two options to fix this are to set it at read and document it (the approach
taken here) or to remove it from the schema (which would be a breaking change,
although not one necessarily with meaningful impact since the value has not been
set at least since the providers were broken out in to their own repositories).