Fix letter case on aws_dms_endpoint.mongodb_settings
AWS DMS automatically lowercases the MongoDB settings values. When Terraform runs, it'll change the settings to uppercase, only to have AWS set them to lowercase. Then, next time, Terraform will notice this change, and try again. This means that, if we set the values as uppercase, the state will never converge. This change fixes that.