website: Link to docs for "attribute-as-blocks mode"
Some initial testing and discussion about the attribute-as-blocks
processing mode revealed that there are some caveats with it that become
apparent in more complex cases. With that in mind, here we link to a
documentation page explaining those caveats to minimize the chance that
users will be surprised by them, and can use the examples on that page
to understand the meaning of the various configuration permutations
supported by these attributes.

This is a catchup PR to add it to the docs for three attributes we already
updated for this mode in previous PRs. Ideally we should add similar links
to the docs for other attributes we'll update in subsequent PRs.