docs/resource/aws_elasticsearch_domain: Add missing equals sign for advanced_options in example
In Terraform 0.12, the equals sign is required as `advanced_options` is an argument. Without running `terraform 0.12upgrade` to upgrade configurations (or if the tool misses this particular configuration), operators will receive a potentially confusing error message on an element of the map, e.g.

```
Error: Invalid argument name

  on main.tf line 60, in resource "aws_elasticsearch_domain" "es":
  60:     "rest.action.multi.allow_explicit_index" = "true"

Argument names must not be quoted.
```

This change is backwards compatible with Terraform 0.11 and prior.