[docs] cloudwatch_metric_alarm
There is a spurious %d in a resource name that makes it seem like it
will use format, and print a number when used with `count`.
This is not the case.

This removes the spurious %d, and adds an example using `format` with
`count`, as I *suspect* this could attribute to issues like:

- https://github.com/terraform-providers/terraform-provider-aws/issues/422

I along with at least one other user, noted that really they had mistakenly
duplicate alarm names. I suspect this could be the case for others as
well.

My hope is that by providing an example to do something like that above,
fewer issues will be opened.