Workaround IAM role creation race with spot instances
Fixes #8164

Due to IAM eventual consistency, it is possible for spot instances to be successfully requested and then
later return `bad-parameters` indicating the IAM role does not exist - but the instance never recovers.

```
<DescribeSpotInstanceRequestsResponse xmlns="http://ec2.amazonaws.com/doc/2016-11-15/">
  <state>failed</state>
  <fault>
      <code>InvalidParameterValue</code>
      <message>Value (dcos-terraform-ci-90913f42df22592c-instance_profile) for parameter iamInstanceProfile.name is invalid. Invalid IAM Instance Profile name</message>
  </fault>
  <status>
      <code>bad-parameters</code>
      <updateTime>2019-04-23T20:22:00.000Z</updateTime>
      <message>Your Spot request failed due to bad parameters.</message>
  </status>
...
}) to resolve: unexpected state 'bad-parameters', wanted target 'fulfilled'. last error: %!s(<nil>)
```

This means that the retries on requesting the spot instance is insufficient - it appears that even if it is successfully
accepted, it can error later. Because of this, this PR makes it so that if `bad-parameters` is returned, we request a new
spot instance and try again.