Not retrieve EMR AD Domain Join User from AWS API
The AWS API returns ******** for the ADDomainJoinUser attribute, the
same as for ADDomainJoinPassword, CrossRealmTrustPrincipalPassword, and
KdcAdminPassword. This causes a terraform plan to always attempt to
rebuild the cluster since the ad_domain_join_user attribute is seen as
changing from ******** to what is specified and desired in the plan.
This now configures EMR to not retrieve this attribute from the AWS APIs
so that it doesn't always attempt to rebuild the cluster.