provider: Hardcode Global Accelerator SDK Region to us-west-2 and Shield SDK Region to us-east-1
References:

* https://docs.aws.amazon.com/general/latest/gr/rande.html#global_accelerator_region
* https://docs.aws.amazon.com/general/latest/gr/rande.html#shield_region

Global AWS services sometimes only use a single region for their endpoint. As a past precedent we have hardcoded the Route 53 SDK Region to us-east-1. Following the current AWS Regions and Endpoints documentation, we do the same type of SDK configuration for Global Accelerator (endpoint is only in us-west-2) and Shield (endpoint is only in us-east-1).