aws_subnet: Increase timeout for delete to 20 mins
We've seen subnet deletion failures that we think we can trace to ENIs left
behind by NLB. We don't have permission to detach or delete those ENIs, so our
only recourse is to wait longer.

Additionally, we plumb through Create and Delete to use the configured timeouts
via `d.Timeout(schema.Timeout{Create,Delete})` so that custom timeouts will be
reflected. Update is not as straightforward to reason about as there are
separate timeouts for disassociation and association of address spaces, and the
individually configured values do not appear to have presented problems so far.