Secure basic CloudTrail example
The problem with the original example is that it sets up a bucket
policy allowing any AWS account to send their CloudTrail logs to the
bucket in question. Perhaps not the most dangerous of attacks, but not
very hygienic either.

The CloudTrail service principal writes to a known bucket path which
includes the origin account id. That is the recommended method to
restrict access.

See also https://docs.aws.amazon.com/awscloudtrail/latest/userguide/create-s3-bucket-policy-for-cloudtrail.html