docs/resource/aws_glue_catalog_table: Use Terraform 0.11 and 0.12 compatible syntax
Reference: https://github.com/terraform-providers/terraform-provider-aws/issues/8877

While the Terraform 0.11 workaround of using a list of maps as an argument to a configuration block attribute did work in certain contexts, its support was unintentional from a configuration language perspective and the argument syntax is explicitly not allowed in Terraform 0.12. This example configuration update is compatible with both Terraform 0.11 and 0.12.