Fix page titles to include full resource/data source name
This changes many documentation files that previously included only part
of the resource name in the page title. For example, the page title for
`aws_cloudwatch_metric_alarm` was `AWS: cloudwatch_metric_alarm` instead
of `AWS: aws_cloudwatch_metric_alarm`.

This was changed to make it easier to search pages from the browser,
searching the history for `aws_cloudwatch_metric` wouldn't show
`cloudwatch_metric_alarm`. Most of the other resources already use the
full name, so it makes sense to standardize on a single form.

Some pages, like for `aws_batch_job_queue`, had even less of the
resource name in the title, like `AWS: batch`.