Clearer source_code_hash description
Since 0.12, wrapping in "${}" is no longer required, so it should be omitted here.

Additionally, you're required to use a `.zip` file for an AWS Lambda, so using `file()` here is not ideal, since you're reading a binary file. Thus, `filebase64()` is a better example, as the ZIP file contents are unlikely to be valid UTF-8 content.