Correct documentation for backup_retention_period and db_subnet_group_name (#8010)
## What

The documentation for `backup_retention_period` is incorrect as the source database should have a greater than 0 value, not the read replica. It was accidentally changed here: https://github.com/terraform-providers/terraform-provider-aws/commit/16424ce60b9edfc44454823d3b0a6f763a007379

Also I clarify for the db_subnet_group_name that `it should be specified only if the source database specifies an instance in another AWS Region`, as if you set it and the source database is in the same region then the AWS API returns an error.