Add validation for all pgp_key arguments
We've had two occurrences recently where `aws_iam_user_login_profile`
resources have failed to apply because the `pgp_key` argument had a key
with an expired or missing subkey.

Make it possible to catch these during code review with `terraform
validate` by attempting to encrypt a dummy value and reporting any
errors. This seems easier and more reliable than attempting to check the
validity of the actual keys and subkeys.

The dummy value is `validation test` because it gets reported in the
error. I could have bypassed this error wrapping by going directly to
`pgpkeys.EncryptShares` but it seemed better to use the same package and
error styling as `encryption.RetrieveGPGKey`.

I've used Hashicorp's key to test the codepath for Keybase because I
expect it to continue to work for the lifetime of this project.

The two keys used by the tests were generated with `gpg --batch
--gen-key` in an `alpine:3.10` container with these two respective
inputs:

    Key-Type: default
    Key-Length: 1024
    Expire-Date: 0
    Name-Real: Without Subkey
    Name-Email: without@sub.key
    %no-protection
    %commit

    Key-Type: default
    Key-Length: 1024
    Subkey-Type: 1
    Subkey-Length: 1024
    Expire-Date: 0
    Name-Real: With Subkey
    Name-Email: with@sub.key
    %no-protection
    %commit

Then they were extracted with `gpg --import <id> | base64`.