Provide a working example for ecs-task-definition
This is working around the issue of not having a task definition when the resources are initially rolled out.

Background:

The documetation example of directly referecing "task_family" doesn't work and exits with an error when initially applying it. See also this issue https://github.com/terraform-providers/terraform-provider-aws/issues/1274

The reason is, that data sources don't handle missing data gracefully. Unfortunately, that's not gonna be addressed, as stated here: https://github.com/hashicorp/terraform/issues/16380#issuecomment-517282340
One of the suggested workarounds is, to add an explict `depends_on`. However, this causes a potential change in the terraform plan output, even though it's not actually going to change. Furthermore, it's discourage by the Terraform documentation itself.

This thread mentions a few other workarounds, but none of them seem to be suitable https://github.com/hashicorp/terraform/issues/16380

`aws_ecs_task_definition.self.revision` can only be referenced, once the resource is created (in contrast to family, which is already present in code)
Apparently, this allows Terraform to correctly resolve the dependencies and makes the data source behave as expected.