deps: Override renovate semanticCommits auto-detection by always disabling
This project contains both non-semantic and semantic commit messages. Renovate tries to auto-detect which style it should use when submitting dependency update pull requests, however the mixture of the two seems to lead to renovate pull requests that start without the `chore(deps)` titles, then flip-flop to including it and not including it. This leads to confusing and duplicate pull requests.

Here we just force disable the semantic commit messaging to prevent the changing auto-detection behavior.

References:

- https://github.com/terraform-providers/terraform-provider-aws/pull/10255 and https://github.com/terraform-providers/terraform-provider-aws/pull/10262
- https://github.com/terraform-providers/terraform-provider-aws/pull/10248 and https://github.com/terraform-providers/terraform-provider-aws/pull/10264
- https://github.com/terraform-providers/terraform-provider-aws/pull/9991 and https://github.com/terraform-providers/terraform-provider-aws/pull/10263
- https://github.com/terraform-providers/terraform-provider-aws/pull/10011#event-2655384146