provider: Remove hashibot pull request labeling of old tags filenames
Reference: https://github.com/terraform-providers/terraform-provider-aws/pull/10325#event-2677341915

Since the file naming and implementation of these files was inconsistent, the labeling was sometimes inconsistent. We won't be accepting changes to these files anymore in preference of replacing them with the auto-generated `keyvaluetags` library, so removing the hashibot labeling of them.