Adds test to confirm current behavior re. iam_role_name
It was my expectation that removing the name from an aws_iam_role
resource would result in terraform generating a new random name for this
resource and forcing a replacement of this resource. This test confirms
that this is not the case currently. Whether this is desired behavior or
not I don't know. I can only add this test in an effort to preserve this
bahavior.