Set `secret` and `ses_smtp_password` to sensitive
According to [article](https://www.terraform.io/docs/extend/best-practices/sensitive-state.html), we should stop focusing efforts on encrypting state values with PGP keys and instead mark fields as sensitive. According to the article, the effort should be spent instead on protecting state buckets and our team intends to move in this direction.

Therefore, I've gone and set both attributes to sensitive and removed the deprecation notice recommending using the `pgp_key` technique.

Since people may disagree, I haven't gone so far as to remove the `pgp_key` support. I think it should remain for as long as the maintainers want to promote people using this feature and as long as it doesn't seem contrarian compared to official Hashicorp doctrine.