resource/aws_appautoscaling_policy: Prevent state removal at creation
References:
* https://github.com/terraform-providers/terraform-provider-aws/issues/10549

The AWS application-autoscaling service has eventual consistency
considerations. The `aws_appautoscaling_policy` resource immediately
tries to read a scaling policy after creation. If the policy is not
found, the application-autoscaling service returns a 200 OK with an
empty list of scaling policies. Since no scaling policy is present, the
`aws_appautoscaling_policy` resource removes the created resource from
state, leading to a "produced an unexpected new value for was present,
but now absent" error.

With the changes in this commit, the empty list of scaling policies in
the response for the newly created resource will result in a
NotFoundError being returned and a retry of the read request. A
subsequent retry should hopefully be successful, leading to the state
being preserved.

Output from acceptance testing:

```
> make testacc TEST=./aws TESTARGS='-run=TestAccAWSAppautoScalingPolicy_'
...
--- PASS: TestAccAWSAppautoScalingPolicy_multiplePoliciesSameName (34.27s)
--- PASS: TestAccAWSAppautoScalingPolicy_dynamoDb (36.59s)
--- PASS: TestAccAWSAppautoScalingPolicy_multiplePoliciesSameResource (37.73s)
--- PASS: TestAccAWSAppautoScalingPolicy_disappears (68.94s)
--- PASS: TestAccAWSAppautoScalingPolicy_basic (72.25s)
--- PASS: TestAccAWSAppautoScalingPolicy_scaleOutAndIn (74.31s)
--- PASS: TestAccAWSAppautoScalingPolicy_spotFleetRequest (77.85s)
--- PASS: TestAccAWSAppautoScalingPolicy_ResourceId_ForceNew (88.97s)
```