tests/resource/aws_eks_fargate_profile: Add PreCheck
Acceptance testing fails in our default region of us-west-2:

```
--- FAIL: TestAccAWSEksFargateProfile_basic (1244.08s)
    testing.go:635: Step 0 error: errors during apply:

        Error: error creating EKS Fargate Profile (tf-acc-test-6753867324954738062:tf-acc-test-6753867324954738062): InvalidRequestException: CreateFargateProfile is not supported for region: us-west-2
```

Determining if this functionality is available in a region requires trying to list or create an EKS Fargate Profile on an existing EKS Cluster, which takes upwards of 15-20 minutes to create and destroy on its own. Rather than introduce an extra 15-20 minutes to see the acceptance test skipping or even continue (then taking its normal 20 minutes), we implement a worst case hardcoded list of allowed regions with information about the list and a link to the most up to date documentation.

Output from acceptance testing in us-west-2:

```
--- SKIP: TestAccAWSEksFargateProfile_basic (2.62s)
    resource_aws_eks_fargate_profile_test.go:280: skipping acceptance testing:

        Test provider region (us-west-2) not found in allowed EKS Fargate regions: [ap-northeast-1 eu-west-1 us-east-1 us-east-2]

        The allowed regions are hardcoded in the acceptance testing since dynamically determining the
        functionality requires creating and destroying a real EKS Cluster, which is a lengthy process.
        If this check is out of date, please consider creating an issue in the Terraform AWS Provider
        repository (https://github.com/terraform-providers/terraform-provider-aws) or updating the
        check itself (testAccPreCheckAWSEksFargateProfile).

        For the most up to date supported region information, see the EKS User Guide:
        https://docs.aws.amazon.com/eks/latest/userguide/fargate.html
```

Output from acceptance testing in us-east-2:

```
--- PASS: TestAccAWSEksFargateProfile_basic (1432.20s)
```