Better debug message in resource_aws_db_option_group
In some cases removing the aws_db_option group will fail complaing that
the option group is still in use, when it is not.

This happens because of a internal snapshot create by AWS
(see issue #4597), and the solution is to wait before trying again.

Feel free to just close this PR if we don't think this is a good idea.

Signed-off-by: fernandoalex <jfernandoalex@gmail.com>