Allow users to set Elastic Beanstalk tags that include `Name`
Elastic Beanstalk automatically creates the `Name` tag for an Elastic
Beanstalk environment, and that tag is not editable.  However, tags
that include `Name` in the tag name are supported by Elastic Beanstalk.

The `aws_elastic_beanstalk_environment` resource incorrectly ignored
any tag that includes `Name` in the tag name.  This updates the regex
used for identifying ignored tags so that only a tag with the exact
name `Name` will be ignored, so that terraform users are not blocked
from creating valid Elastic Beanstalk environment tags.

    https://docs.aws.amazon.com/elasticbeanstalk/latest/dg/using-features.tagging.html