[RM-3123] Remove ConflictsWith for launch template security groups
It should not be possible to set both `security_group_names` and
`vpc_security_group_ids` at the same time, but someone managed to do it anyway.

It is important to note that the contents of the two fields is the same;
except the `security_group_names` is a bit bogus so I suspect there is some
AWS bug that happened during validation and allowed this to be set.

This means that in cases where both are set; we can assume they are set
to the same value and for drift we only want to set one of them; which is
`vpc_security_group_ids`, since `security_group_names` seems deprecated in
favor of that.