aws/resource_aws_route53_record.go: update change record set to use SDK backoff
`PriorRequestNotComplete` is already part of the list of throttle codes [1], so if we use the SDK inbuilt backoff we shouldn't need custom state change logic that retries for 5 minutes.

So this changes the changeRoute53RecordSet to resource.Retry and last retryOnTimeout from issue 7873 [2], such that we send request in resource.Retry for 1 minute and if there is a timeout, we send another request that then uses the
AWS SDK default retry logic...

while the deleteRoute53RecordSet now directly makes the request using SDK without any external retries.

[1]: https://github.com/aws/aws-sdk-go/blob/3acad1210f5c4feb4866ff9874b5f3f5895bd96d/aws/request/retryer.go#L93
[2]: https://github.com/terraform-providers/terraform-provider-aws/issues/7873