Fix wait for ASG capacity when increasing min size
Fixes #5241.

When updating an ASG there are currently two ways
`capacitySatisfiedUpdate()` might ultimately be called from
`resourceAwsAutoscalingGroupUpdate()`:

 - `if d.HasChange("min_size")`, and/or
 - `if d.HasChange("desired_capacity")`

The existing approach here, only checking against `desired_capacity`,
is flawed because that argument is optional while `min_size` is
mandatory. If `desired_capacity` is not specified, AWS will
automatically [update it](https://docs.aws.amazon.com/autoscaling/ec2/APIReference/API_UpdateAutoScalingGroup.html)
to match either `min_size` or `max_size` depending on whether the group
is being increased or decreased.

However, the Terraform AWS provider does not predict this behaviour in
advance. The new auto-updated value for `desired_capacity` is only
detected at the very end of `resourceAwsAutoscalingGroupUpdate()`
(after waiting for capacity) when the configuration of the updated ASG
is read. Before then `d.Get("desired_capacity")` returns the existing
value.

All of this presents a problem when `desired_capacity` is not specified
and `min_size` is increased. We detect a change to `min_size` and thus
trigger a `waitForASGCapacity()`, but then wait for exactly (existing)
`desired_capacity` healthy instances. This number is less than both
`min_size` and the new `desired_capacity`, meaning that even if the
update is successful, we get stuck forever waiting for a smaller number
of instances.

The fix is to wait for whichever is the higher of `min_size` and
`desired_capacity` instances.