resource/aws_backup_selection: Retry for IAM policy errors
References:
* https://github.com/terraform-providers/terraform-provider-aws/issues/10511

The AWS backup service has eventual consistency considerations. In #9298
(d799998c), some logic was added to retry for CreateBackupSelection
calls which failed with an error that the associated IAM role "cannot be
assumed". Occasionally, a similar error appears when the IAM role "is
not authorized to call tag:GetResources". This commit adds logic to
retry for the "is not authorized to call" errors.

Output from acceptance testing:

```
make testacc TEST=./aws TESTARGS='-run=TestAccAwsBackupSelection_'
...
--- PASS: TestAccAwsBackupSelection_disappears (33.19s)
--- PASS: TestAccAwsBackupSelection_withTags (33.56s)
--- PASS: TestAccAwsBackupSelection_basic (34.41s)
--- PASS: TestAccAwsBackupSelection_withResources (43.25s)
--- PASS: TestAccAwsBackupSelection_updateTag (52.79s)
```