Update ecs_service.html.markdown
Confirmed that the service does not recognize the capacity provider when the ecs service is pointed to the arn of the capacity provider. This returns the error message "No Container Instances were found in your capacity provider" and the Cloudwatch metric  "CapacityProviderReservation" is not adjusted. Once the name is used for the capacity provider the service properly launches a task in the provisioning state, "CapacityProviderReservation" is incremented, and an instance is added to the ASG.