Update 'aws_route' to handle changes in IPv6 route entries
'aws_route' allows you to create route table entries with IPv6
destinations and targets including (but not limited to) a gateway,
an EC2 instance, a network interface, a transit_gateway or a
vpc_peering_connection. However, none of the entries with those
targets could *update* their target, since the update function
assumed that all updates would be IPv4 entries.