resource/aws_lambda_permission: Fixes for tfproviderlint R006 (#12044)
Reference: https://github.com/terraform-providers/terraform-provider-aws/issues/11864

`RetryFunc` should only be used when logic has a retryable condition. In the case of working with the AWS Go SDK, it also arbitrarily restricts the automatic retrying logic of API calls to the timeout, which is generally undesired.

In this particular case, the code had no retryable conditions, so the retry logic was extraneous. It is worth noting that this type of delete-then-check code in a Terraform resource is very uncommon and likely should be removed, although there is a note at the top of the deletion function that mentions an acknowledged API bug with parallel requests so leaving this in here for now.

Previously:

```
aws/resource_aws_lambda_permission.go:339:38: R006: RetryFunc should include RetryableError() handling or be removed
```

Output from acceptance testing:

```
--- PASS: TestAccAWSLambdaPermission_withS3 (55.05s)
--- PASS: TestAccAWSLambdaPermission_withSNS (60.76s)
--- PASS: TestAccAWSLambdaPermission_basic (67.48s)
--- PASS: TestAccAWSLambdaPermission_withRawFunctionName (79.46s)
--- PASS: TestAccAWSLambdaPermission_withIAMRole (86.26s)
--- PASS: TestAccAWSLambdaPermission_multiplePerms (92.22s)
--- PASS: TestAccAWSLambdaPermission_withStatementIdPrefix (93.41s)
--- PASS: TestAccAWSLambdaPermission_withQualifier (100.65s)
--- PASS: TestAccAWSLambdaPermission_StatementId_Duplicate (102.96s)
```