resource/aws_route53_health_check: normalize ipv6 addresses in healthchecks
Without this change, terraform will generate a recreate plan each time
if the resource is specified with an ipv6 address that differs from how
aws stores it.

This includes recreating if the same configuration, using a
non-aws-canonical ipv6 form, is applied twice in a row.

Fix this by normalizing the value to whatever go's `net.IP.String()`
formats it as.