Update iam_role.html.markdown
Correct the documentation to identify that it is required to remove all policies before destroying a role and not just the ones that Terraform attached.  This should not even be an option and if the intention is to destroy a role, performing a destruction should just simply remove all policies then destroy the role.  I am truly confused as to why this option even exists.