internal/keyvaluetags: Initial tagging function generator for handling tagging of new resources
Some services require special handling for newly created resources, such as retrying on "not found" errors for eventual consistency. This generator starts with just the EC2 service implementation, but is extensible for any service.

We can also consider generating this for all services to simplify `{SERVICE}UpdateTags(conn, id, nil, tags)` usage, but that is not a current refactoring goal.