r/aws_route53_health_check: Validate reference_name
Support plan-time validation for the length of the reference_name
argument. Since this resource auto-appends a unique string to the
provided argument, this is necessary to avoid obtuse errors on apply.
This validates the provided reference_name plus the generated string is
no longer than 64 chars, the actual API limit.

Since the generated unique ID is 37 chars, the `reference_name` argument
cannot exceed 27 characters.

Closes #7248

Co-Authored-By: Brian Flad <bflad417@gmail.com>