infrastructure/repository: Lift and shift Terraform configuration for managing certain repository labels (#13212)
Migrating from `terraform-repositories`, where it was already separately handled. Other more sensitive repository management and labeling will remain there. After this is merged, will update Terraform Cloud VCS to point at this repository and path so it can automatically trigger runs on the default branch and pull requests when this directory is updated.